basis.router: renamed spec file name for ParametrizedRoute
