+++ test/index.js
@@ -4,30 +4,30 @@ module.exports = {
   html: __dirname + '/env.html', // base env
   test: [
     //require('./spec/common/classes.js'),
-    // require('./spec/core.js'),
-    // require('./spec/date.js'),
-    // require('./spec/crypt.js'),
-    // require('./spec/ua.js'),
+    require('./spec/core.js'),
+    require('./spec/date.js'),
+    require('./spec/crypt.js'),
+    require('./spec/ua.js'),
     require('./spec/router.js'),
-    require('./spec/router/new.js'),
-    // require('./spec/l10n.js'),
-    // require('./spec/data.js'),
-    // require('./spec/data.object.js'),
-    // require('./spec/data.dataset.js'),
-    // require('./spec/data.value.js'),
-    // require('./spec/data.index.js'),
-    // require('./spec/data.vector.js'),
-    // require('./spec/entity.js'),
-    // require('./spec/dom.wrapper.js'),
-    // require('./spec/template.js'),
-    // require('./spec/ui.js'),
-    // require('./spec/net.js'),
-    // require('./spec/dom.js'),
-    // require('./spec/cssom.js'),
-    // require('./spec/layout.js'),
-    // require('./spec/utils.info.js'),
-    // require('./spec/tracker.js'),
-    // require('./spec/type.js')
+    require('./spec/router/parametrized.js'),
+    require('./spec/l10n.js'),
+    require('./spec/data.js'),
+    require('./spec/data.object.js'),
+    require('./spec/data.dataset.js'),
+    require('./spec/data.value.js'),
+    require('./spec/data.index.js'),
+    require('./spec/data.vector.js'),
+    require('./spec/entity.js'),
+    require('./spec/dom.wrapper.js'),
+    require('./spec/template.js'),
+    require('./spec/ui.js'),
+    require('./spec/net.js'),
+    require('./spec/dom.js'),
+    require('./spec/cssom.js'),
+    require('./spec/layout.js'),
+    require('./spec/utils.info.js'),
+    require('./spec/tracker.js'),
+    require('./spec/type.js')
   ]
 };
